refactored(utils/utils):
relocated getter logic to controller
relocated try catch to handler
