+++ PF_Api/src/handlers/productHandlers.js
@@ -1,7 +1,36 @@
-const {productCreator, getAProduct, getProducts} = require("../controllers/productControllers.js");
+const {productCreator, getAProduct, getProducts, getProduct} = require("../controllers/productControllers.js");
 
 module.exports = {
-  postProductHandler: async (req, res) => {},
-  getProductsHandler: async (req, res) => {},
-  getAProductHandler: async (req, res) => {},
+  postProductHandler: async (req, res) => {
+    try {
+      
+    } catch (error) {
+      
+    }
+  },
+  getProductsHandler: async (req, res) => {
+    try {
+      
+    } catch (error) {
+      
+    }
+  },
+  getAProductHandler: async (req, res) => {
+    const {name} = req.body;
+    try {
+      if (name) {
+        // obtener 1 producto por nombre
+        const prod = await getProduct(name);
+        res.status(200).json(prod)
+      } else {
+        console.log("GET all dogs request");
+        // obtener todos los productos
+        const allProducts = await getProducts();
+        res.status(200).json(allProducts)
+      }   
+    } catch (error) {
+      console.log(error.message);
+      res.status(401).send("Data API error, not found");
+    }
+  },
 };
